+++ NEWS.md
@@ -1,3 +1,14 @@
+sport v0.1.2 (Release date: 2019-01-06)
+==============
+
+Changes:
+
+* `sigma` argument replaced with `lambda`.
+* removed `beta` from glicko `gamma` from bbt.
+* all algorithms with players nested in teams.
+* additional `share` argument for `player | team` events. 
+* increased speed of the algorithms by rewriting event loop to c++.
+
 sport v0.1.2 (Release date: 2018-01-07)
 ==============